+++ ui-integration-test/src/main/java/co/electriccoin/zcash/ui/integration/test/screen/scan/view/ScanViewTestSetup.kt
@@ -4,7 +4,7 @@ import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import cash.z.ecc.android.sdk.type.AddressType
-import co.electriccoin.zcash.ui.common.model.WalletRestoringState
+import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.integration.test.common.getPermissionNegativeButtonUiObject
 import co.electriccoin.zcash.ui.integration.test.common.getPermissionPositiveButtonUiObject
@@ -59,7 +59,7 @@ class ScanViewTestSetup(
             onScanStateChanged = {
                 scanState.set(it)
             },
-            walletRestoringState = WalletRestoringState.NONE,
+            topAppBarSubTitleState = TopAppBarSubTitleState.None,
             addressValidationResult = AddressType.Unified
         )
     }