+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/model/TopAppBarSubTitleState.kt
@@ -1,0 +1,9 @@
+package co.electriccoin.zcash.ui.common.model
+
+sealed class TopAppBarSubTitleState {
+    data object None : TopAppBarSubTitleState()
+
+    data object Disconnected : TopAppBarSubTitleState()
+
+    data object Restoring : TopAppBarSubTitleState()
+}